backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
weaviate-client>=4.0.0
sentence-transformers
numpy
openai
